requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.1
django-appconf==1.0.6
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-js-asset==2.2.0
idna==3.6
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.4
urllib3==2.1.0
